Resolv requests changes

+++ to/controlers/create_trajectory.cpp
@@ -8,7 +8,7 @@
 
 // Created by: Omar Nassef (KCL)
 
-#include "include/create_trajectory.h"
+#include "create_trajectory.h"
 
 #include <iostream>
 #include <vector>
@@ -134,10 +134,9 @@ std::optional<vector<float>> RoadUsertoModelInput(const std::shared_ptr<RoadUser
         }
     }
 
-		if (!x) {
-				return std::nullopt;
-		}
-
+    if (!x) {
+            return std::nullopt;
+    }
 
     std::vector<float> mergingCar;
     mergingCar.push_back(RoadUserGPStoProcessedGPS(merging_car->getLongitude()));
@@ -195,13 +194,13 @@ auto calculatedTrajectories(const std::shared_ptr<Database>& database,
     auto calculated_n_1_states = valid_action.first;
 
     auto waypoint{std::make_shared<Waypoint>()};
-    waypoint->setTimestamp(timestamp + (distanceEarth(RoadUserGPStoProcessedGPS(mergingVehicle->getLatitude()), RoadUserGPStoProcessedGPS(mergingVehicle->getLongitude()),
-                  calculated_n_1_states[0][0].item<float>(),calculated_n_1_states[0][1].item<float>()) / fmax(calculated_n_1_states[0][4].item<float>(),float(70))) * 100); //distance to mergeing point
-    waypoint->setLongitude(ProcessedGPStoRoadUserGPS(calculated_n_1_states[0][0].item<float>()));
-    waypoint->setLatitude(ProcessedGPStoRoadUserGPS(calculated_n_1_states[0][1].item<float>()));
-    waypoint->setSpeed(ProcessedSpeedtoRoadUserSpeed(fmax(calculated_n_1_states[0][4].item<float>(),float(70)) / 3.6));
+    waypoint->setTimestamp(static_cast<uint64_t>(timestamp + (distanceEarth(RoadUserGPStoProcessedGPS(mergingVehicle->getLatitude()), RoadUserGPStoProcessedGPS(mergingVehicle->getLongitude()),
+                  calculated_n_1_states[0][0].item<float>(),calculated_n_1_states[0][1].item<float>()) / fmax(calculated_n_1_states[0][4].item<float>(),float(70))) * 100)); //distance to mergeing point
+    waypoint->setLongitude(static_cast<uint32_t>(ProcessedGPStoRoadUserGPS(calculated_n_1_states[0][0].item<float>())));
+    waypoint->setLatitude(static_cast<uint32_t>(ProcessedGPStoRoadUserGPS(calculated_n_1_states[0][1].item<float>())));
+    waypoint->setSpeed(ProcessedSpeedtoRoadUserSpeed(static_cast<int>(fmax(calculated_n_1_states[0][4].item<float>(),float(70)) / 3.6)));
     (!(valid_action.second)) ? waypoint->setLanePosition(mergingVehicle->getLanePosition() + 1) : waypoint->setLanePosition(mergingVehicle->getLanePosition());
-		waypoint->setHeading(ProcessedHeadingtoRoadUserHeading(calculated_n_1_states[0][19].item<float>()));
+		waypoint->setHeading(static_cast<uint16_t>(ProcessedHeadingtoRoadUserHeading(calculated_n_1_states[0][19].item<float>())));
     mergingManeuver->addWaypoint(waypoint);
 		mergingVehicle->setProcessingWaypoint(true);
